Exercise 13. ContactDetailsTests; Carrying out a test checking the website with detailed contact details.
If during the creation of a new contact I select a group of contacts in the Groups field, information about the group is on the page with contact information and there is no such field on the contact modification page, so the ContactDetailsTests test falls. To solve this problem, at the moment I do not complete the information about the group when creating a new contact (I select the group in the "none" field). Then the entire test packet works correctly.